# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MVTable 后端服务 API",
        "title": "MVTable",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MVTable",
            "email": "mvtable@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/project/advanced-perm/disable": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "关闭项目高级权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目高级权限"
                ],
                "summary": "关闭项目高级权限",
                "parameters": [
                    {
                        "description": "关闭项目高级权限请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DisableMvProjectAdvancedPermReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "关闭成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/project/advanced-perm/enable": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "开启项目高级权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目高级权限"
                ],
                "summary": "开启项目高级权限",
                "parameters": [
                    {
                        "description": "开启项目高级权限请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EnableMvProjectAdvancedPermReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "开启成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/project/advanced-perm/get": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取项目高级权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目高级权限"
                ],
                "summary": "获取项目高级权限",
                "parameters": [
                    {
                        "type": "string",
                        "name": "role",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "tableSchemaId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.GetMvProjectAdvancedPermRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/project/advanced-perm/update": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新项目高级权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目高级权限"
                ],
                "summary": "更新项目高级权限",
                "parameters": [
                    {
                        "description": "更新项目高级权限请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateMvProjectAdvancedPermReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/view/table/update": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新MV视图表格信息\nfilterConfig: [{fieldId: \"字段id\", operator: \"equal=等于/notEqual=不等于/contains=包含/notContains=不包含/null=为空/notNull=不为空/before=早于/after=晚于/beforeOrEqual=早于或等于/afterOrEqual=晚于或等于\", value: \"值\"}]\ngroupConfig: {fieldId: \"\", order: \"asc/desc\"}\nsortConfig: [{fieldId: \"\", order: \"asc/desc\"}]\nrowHeight: 1=常规, 2=中等, 3=高, 4=超高\ncolorConfig: [{color: \"#cccccc\", scope: \"cell/row/rowHeader/column\", fieldId: \"字段id\", operator: \"equal=等于/notEqual=不等于/greaterThan=大于/lessThan=小于/greaterThanOrEqual=大于等于/lessThanOrEqual=小于等于/contains=包含/notContains=不包含/null=为空/notNull=不为空/before=早于/after=晚于/beforeOrEqual=早于或等于/afterOrEqual=晚于或等于\", value: \"值\"}]",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "视图表格"
                ],
                "summary": "更新视图表格",
                "parameters": [
                    {
                        "description": "更新视图表格请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateMvViewTableReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "middleware.Response": {
            "description": "统一响应结构",
            "type": "object",
            "properties": {
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string"
                },
                "state": {
                    "description": "响应状态码",
                    "type": "integer"
                }
            }
        },
...
        "model.DisableMvProjectAdvancedPermReq": {
            "type": "object",
            "properties": {
                "projectId": {
                    "type": "string"
                }
            }
        },
        "model.EnableMvProjectAdvancedPermReq": {
            "type": "object",
            "properties": {
                "projectId": {
                    "type": "string"
                }
            }
        },
        "model.FieldPerm": {
            "type": "object",
            "properties": {
                "canAdd": {
                    "type": "boolean"
                },
                "canEdit": {
                    "type": "boolean"
                },
                "canRead": {
                    "type": "boolean"
                },
                "fieldId": {
                    "type": "string"
                }
            }
        },
...
        "model.GetMvProjectAdvancedPermRes": {
            "type": "object",
            "properties": {
                "canAdd": {
                    "type": "boolean"
                },
                "canCheckViews": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "canDelete": {
                    "type": "boolean"
                },
                "customFieldPerm": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldPerm"
                    }
                },
                "dataAction": {
                    "type": "string"
                },
                "fieldAccess": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "operateRange": {
                    "type": "string"
                },
                "viewAccess": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateMvProjectAdvancedPermReq": {
            "type": "object",
            "required": [
                "dataAction"
            ],
            "properties": {
                "canAdd": {
                    "type": "boolean"
                },
                "canCheckViews": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "canDelete": {
                    "type": "boolean"
                },
                "canOperateView": {
                    "type": "boolean"
                },
                "customFieldPerm": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldPerm"
                    }
                },
                "dataAction": {
                    "type": "string"
                },
                "fieldAccess": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "operateRange": {
                    "type": "string"
                },
                "viewAccess": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateMvViewTableReq": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "输入Bearer token进行认证，格式: Bearer {token}。例如: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}